ecommerce-automation/requirements.txt
playwright==1.40.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
beautifulsoup4==4.12.2
selenium==4.15.0 
undetected-chromedriver==3.5.4
Pillow==10.1.0
asyncio-throttle==1.0.2